+++ lib/src/configuration/configuration_spanning_tree.dart
@@ -8,4 +8,12 @@ class ConfigurationSpanningTree {
 
   void enableRapidPVST() => device
     ..run('spanning-tree mode rapid-pvst');
+
+  void setPriorityValue(int vlan, int priority) {
+    if (priority % 4096 != 0 || 15 < priority / 4096) {
+      throw ArgumentError.value(priority);
+    }
+
+    device.run('spanning-tree vlan $vlan priority $priority');
+  }
 }